Task Diagram Implementation for Milestone 2
1 March 2016 -
- Full skeleton code complete, task diagram implementation
- Control thread receives from UART, turns on/off LED
- Control thread receives from UART, sends to motor thread, UART sends
response from motor thread
- Sensor thread gets ADC, sends to control, sends over UART

+++ Embedded/firmware/src/debug.c
@@ -1,6 +1,6 @@
 #include "debug.h"
 
-void outputEvent( uint32_t value  )
+void outputEvent( uint8_t value  )
 {    
     //Get current value from Port E
     uint32_t portE = SYS_PORTS_Read(PORTS_ID_0, PORT_CHANNEL_E);
